# docs/openapi/spec.yaml
openapi: 3.1.0
info:
  title: Eazy Recycling Backend Service
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
paths:
  /weight-tickets/{weightTicketId}:
    get:
      tags:
      - weight-ticket-controller
      operationId: getWeightTicketByNumber
      parameters:
      - name: weightTicketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/WeightTicketDetailView"
    put:
      tags:
      - weight-ticket-controller
      operationId: update
      parameters:
      - name: weightTicketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightTicketRequest"
        required: true
      responses:
        "204":
          description: No Content
...
  /invoices/{id}:
    get:
      tags:
      - invoice-controller
      operationId: getById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InvoiceDetailView"
    put:
      tags:
      - invoice-controller
      operationId: update_2
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInvoiceRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InvoiceResult"
    delete:
      tags:
      - invoice-controller
      operationId: delete
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
...
  /weight-tickets:
    get:
      tags:
      - weight-ticket-controller
      operationId: getWeightTickets
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeightTicketListView"
    post:
      tags:
      - weight-ticket-controller
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightTicketRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateWeightTicketResponse"
  /weight-tickets/{weightTicketId}/split:
    post:
      tags:
      - weight-ticket-controller
      operationId: split
      parameters:
      - name: weightTicketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SplitWeightTicketRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/SplitWeightTicketResponse"
  /weight-tickets/{weightTicketId}/invoice:
    post:
      tags:
      - weight-ticket-controller
      operationId: createInvoice
      parameters:
      - name: weightTicketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateInvoiceFromWeightTicketResult"
  /weight-tickets/{weightTicketId}/copy:
    post:
      tags:
      - weight-ticket-controller
      operationId: copy
      parameters:
      - name: weightTicketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CopyWeightTicketResponse"
  /weight-tickets/{weightTicketId}/complete:
    post:
      tags:
      - weight-ticket-controller
      operationId: complete
      parameters:
      - name: weightTicketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: No Content
  /weight-tickets/{weightTicketId}/cancel:
    post:
      tags:
      - weight-ticket-controller
      operationId: cancel
      parameters:
      - name: weightTicketId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelWeightTicketRequest"
        required: true
      responses:
        "204":
          description: No Content
  /weight-tickets/completed:
    post:
      tags:
      - weight-ticket-controller
      operationId: createCompleted
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WeightTicketRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateWeightTicketResponse"
...
  /invoices:
    get:
      tags:
      - invoice-controller
      operationId: getAll
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InvoiceView"
    post:
      tags:
      - invoice-controller
      operationId: create_2
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InvoiceResult"
  /invoices/{id}/finalize:
    post:
      tags:
      - invoice-controller
      operationId: finalize
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InvoiceResult"
  /invoices/completed:
    post:
      tags:
      - invoice-controller
      operationId: createCompleted_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/InvoiceResult"
...
components:
  schemas:
    CompanyConsignor:
      allOf:
      - $ref: "#/components/schemas/ConsignorRequest"
      - type: object
        properties:
          companyId:
            type: string
            format: uuid
      required:
      - companyId
...
    NoPickupLocationRequest:
      allOf:
      - $ref: "#/components/schemas/PickupLocationRequest"
    PersonConsignor:
      allOf:
      - $ref: "#/components/schemas/ConsignorRequest"
    PickupCompanyRequest:
      allOf:
      - $ref: "#/components/schemas/PickupLocationRequest"
      - type: object
        properties:
          companyId:
            type: string
            format: uuid
      required:
      - companyId
...
    WeightTicketLineRequest:
      type: object
      properties:
        wasteStreamNumber:
          type: string
        weight:
          $ref: "#/components/schemas/WeightRequest"
        catalogItemId:
          type: string
          format: uuid
      required:
      - catalogItemId
      - weight
    WeightTicketRequest:
      type: object
      properties:
        consignorParty:
          oneOf:
          - $ref: "#/components/schemas/CompanyConsignor"
          - $ref: "#/components/schemas/PersonConsignor"
        lines:
          type: array
          items:
            $ref: "#/components/schemas/WeightTicketLineRequest"
        tarraWeightValue:
          type: string
        tarraWeightUnit:
          type: string
          enum:
          - KG
        secondWeighingValue:
          type: string
        secondWeighingUnit:
          type: string
          enum:
          - KG
        weightedAt:
          type: string
          format: date
        carrierParty:
          type: string
          format: uuid
        direction:
          type: string
          enum:
          - INBOUND
          - OUTBOUND
        pickupLocation:
          oneOf:
          - $ref: "#/components/schemas/DutchAddressRequest"
          - $ref: "#/components/schemas/NoPickupLocationRequest"
          - $ref: "#/components/schemas/PickupCompanyRequest"
          - $ref: "#/components/schemas/ProjectLocationRequest"
          - $ref: "#/components/schemas/ProximityDescriptionRequest"
        deliveryLocation:
          oneOf:
          - $ref: "#/components/schemas/DutchAddressRequest"
          - $ref: "#/components/schemas/NoPickupLocationRequest"
          - $ref: "#/components/schemas/PickupCompanyRequest"
          - $ref: "#/components/schemas/ProjectLocationRequest"
          - $ref: "#/components/schemas/ProximityDescriptionRequest"
        truckLicensePlate:
          type: string
        reclamation:
          type: string
        note:
          type: string
      required:
      - consignorParty
      - direction
      - lines
...
    UpdateInvoiceRequest:
      type: object
      properties:
        invoiceDate:
          type: string
          format: date
        lines:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineRequest"
      required:
      - invoiceDate
      - lines
    InvoiceResult:
      type: object
      properties:
        invoiceId:
          type: string
          format: uuid
      required:
      - invoiceId
...
    CompanyView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        chamberOfCommerceId:
          type: string
        vihbId:
          type: string
        processorId:
          type: string
        address:
          $ref: "#/components/schemas/AddressView"
        email:
          type: string
      required:
      - address
      - id
      - name
    DutchAddressView:
      allOf:
      - $ref: "#/components/schemas/PickupLocationView"
      - type: object
        properties:
          streetName:
            type: string
          postalCode:
            type: string
          buildingNumber:
            type: string
          buildingNumberAddition:
            type: string
          city:
            type: string
          country:
            type: string
      required:
      - buildingNumber
      - city
      - country
      - postalCode
      - streetName
    NoPickupView:
      allOf:
      - $ref: "#/components/schemas/PickupLocationView"
    PickupCompanyView:
      allOf:
      - $ref: "#/components/schemas/PickupLocationView"
      - type: object
        properties:
          company:
            $ref: "#/components/schemas/CompanyView"
      required:
      - company
...
    ProjectLocationView:
      allOf:
      - $ref: "#/components/schemas/PickupLocationView"
      - type: object
        properties:
          id:
            type: string
          company:
            $ref: "#/components/schemas/CompanyView"
          streetName:
            type: string
          postalCode:
            type: string
          buildingNumber:
            type: string
          buildingNumberAddition:
            type: string
          city:
            type: string
          country:
            type: string
      required:
      - buildingNumber
      - city
      - company
      - country
      - id
      - postalCode
      - streetName
    ProximityDescriptionView:
      allOf:
      - $ref: "#/components/schemas/PickupLocationView"
      - type: object
        properties:
          postalCodeDigits:
            type: string
          city:
            type: string
          description:
            type: string
          country:
            type: string
      required:
      - city
      - country
      - description
      - postalCodeDigits
...
    CreateWeightTicketResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
      required:
      - id
    SplitWeightTicketRequest:
      type: object
      properties:
        originalWeightTicketPercentage:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
        newWeightTicketPercentage:
          type: integer
          format: int32
          maximum: 100
          minimum: 1
      required:
      - newWeightTicketPercentage
      - originalWeightTicketPercentage
    SplitWeightTicketResponse:
      type: object
      properties:
        originalWeightTicketId:
          type: integer
          format: int64
        newWeightTicketId:
          type: integer
          format: int64
      required:
      - newWeightTicketId
      - originalWeightTicketId
    CreateInvoiceFromWeightTicketResult:
      type: object
      properties:
        invoiceId:
          type: string
          format: uuid
        weightTicketId:
          type: integer
          format: int64
      required:
      - invoiceId
      - weightTicketId
    CopyWeightTicketResponse:
      type: object
      properties:
        originalWeightTicketId:
          type: integer
          format: int64
        newWeightTicketId:
          type: integer
          format: int64
      required:
      - newWeightTicketId
      - originalWeightTicketId
    CancelWeightTicketRequest:
      type: object
      properties:
        cancellationReason:
          type: string
          minLength: 1
      required:
      - cancellationReason
...
    CreateInvoiceRequest:
      type: object
      properties:
        invoiceType:
          type: string
          minLength: 1
        documentType:
          type: string
          minLength: 1
        customerId:
          type: string
          minLength: 1
        invoiceDate:
          type: string
          format: date
        originalInvoiceId:
          type: string
          format: uuid
        lines:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineRequest"
      required:
      - customerId
      - documentType
      - invoiceDate
      - invoiceType
      - lines
...
    WeightTicketListView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        consignorPartyName:
          type: string
        totalWeight:
          type: number
          format: double
        weighingDate:
          type: string
          format: date-time
        note:
          type: string
        status:
          type: string
          enum:
          - DRAFT
          - COMPLETED
          - INVOICED
          - CANCELLED
      required:
      - consignorPartyName
      - id
      - status
    CompanyConsignorView:
      allOf:
      - $ref: "#/components/schemas/ConsignorView"
      - type: object
        properties:
          company:
            $ref: "#/components/schemas/CompanyView"
      required:
      - company
    ConsignorView:
      discriminator:
        propertyName: type
      properties:
        type:
          type: string
      required:
      - type
...
    PersonConsignorView:
      allOf:
      - $ref: "#/components/schemas/ConsignorView"
    WeightTicketDetailView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        consignorParty:
          oneOf:
          - $ref: "#/components/schemas/CompanyConsignorView"
          - $ref: "#/components/schemas/PersonConsignorView"
        status:
          type: string
        lines:
          type: array
          items:
            $ref: "#/components/schemas/WeightTicketLineView"
        secondWeighingValue:
          type: number
        secondWeighingUnit:
          type: string
        tarraWeightValue:
          type: number
        tarraWeightUnit:
          type: string
        carrierParty:
          $ref: "#/components/schemas/CompanyView"
        direction:
          type: string
        pickupLocation:
          oneOf:
          - $ref: "#/components/schemas/DutchAddressView"
          - $ref: "#/components/schemas/NoPickupView"
          - $ref: "#/components/schemas/PickupCompanyView"
          - $ref: "#/components/schemas/ProjectLocationView"
          - $ref: "#/components/schemas/ProximityDescriptionView"
        deliveryLocation:
          oneOf:
          - $ref: "#/components/schemas/DutchAddressView"
          - $ref: "#/components/schemas/NoPickupView"
          - $ref: "#/components/schemas/PickupCompanyView"
          - $ref: "#/components/schemas/ProjectLocationView"
          - $ref: "#/components/schemas/ProximityDescriptionView"
        truckLicensePlate:
          type: string
        reclamation:
          type: string
        note:
          type: string
        cancellationReason:
          type: string
        linkedInvoiceId:
          type: string
          format: uuid
        createdAt:
          $ref: "#/components/schemas/LocalDateTime"
        createdByName:
          type: string
        updatedAt:
          $ref: "#/components/schemas/LocalDateTime"
        updatedByName:
          type: string
        weightedAt:
          $ref: "#/components/schemas/LocalDateTime"
        pdfUrl:
          type: string
      required:
      - consignorParty
      - direction
      - id
      - lines
      - status
    WeightTicketLineView:
      type: object
      properties:
        wasteStreamNumber:
          type: string
        catalogItemId:
          type: string
          format: uuid
        weightValue:
          type: number
        weightUnit:
          type: string
      required:
      - catalogItemId
      - weightUnit
      - weightValue
...
    InvoiceView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        invoiceType:
          type: string
        documentType:
          type: string
        status:
          type: string
        invoiceDate:
          type: string
          format: date
        customerName:
          type: string
        customerNumber:
          type: string
        totalExclVat:
          type: number
        totalInclVat:
          type: number
      required:
      - customerName
      - documentType
      - id
      - invoiceDate
      - invoiceType
      - status
      - totalExclVat
      - totalInclVat
...
    InvoiceDetailView:
      type: object
      properties:
        id:
          type: string
          format: uuid
        invoiceNumber:
          type: string
        invoiceType:
          type: string
        documentType:
          type: string
        status:
          type: string
        invoiceDate:
          type: string
          format: date
        customer:
          $ref: "#/components/schemas/CustomerSnapshotView"
        originalInvoiceId:
          type: string
          format: uuid
        sourceWeightTicketId:
          type: integer
          format: int64
        lines:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceLineView"
        totals:
          $ref: "#/components/schemas/InvoiceTotalsView"
        createdAt:
          type: string
        createdBy:
          type: string
        updatedAt:
          type: string
        updatedBy:
          type: string
        finalizedAt:
          type: string
        finalizedBy:
          type: string
        pdfUrl:
          type: string
      required:
      - createdAt
      - customer
      - documentType
      - id
      - invoiceDate
      - invoiceType
      - lines
      - status
      - totals
...
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT